hospital-agent/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
numpy
PyJWT
cryptography